refactor: move cards into separate dir
