+++ catalyst_voices/apps/voices/lib/pages/dev_tools/cards/logs_card.dart
@@ -1,4 +1,4 @@
-import 'package:catalyst_voices/pages/dev_tools/widgets/info_card.dart';
+import 'package:catalyst_voices/pages/dev_tools/cards/info_card.dart';
 import 'package:catalyst_voices/pages/dev_tools/widgets/logs_collect_toggle.dart';
 import 'package:catalyst_voices/pages/dev_tools/widgets/logs_export_button.dart';
 import 'package:catalyst_voices/pages/dev_tools/widgets/logs_level_selector.dart';
