+++ catalyst_voices/apps/voices/lib/pages/dev_tools/dev_tools_page.dart
@@ -1,8 +1,8 @@
-import 'package:catalyst_voices/pages/dev_tools/widgets/app_info_card.dart';
-import 'package:catalyst_voices/pages/dev_tools/widgets/config_card.dart';
-import 'package:catalyst_voices/pages/dev_tools/widgets/documents_card.dart';
-import 'package:catalyst_voices/pages/dev_tools/widgets/gateway_info_card.dart';
-import 'package:catalyst_voices/pages/dev_tools/widgets/logs_card.dart';
+import 'package:catalyst_voices/pages/dev_tools/cards/app_info_card.dart';
+import 'package:catalyst_voices/pages/dev_tools/cards/config_card.dart';
+import 'package:catalyst_voices/pages/dev_tools/cards/documents_card.dart';
+import 'package:catalyst_voices/pages/dev_tools/cards/gateway_info_card.dart';
+import 'package:catalyst_voices/pages/dev_tools/cards/logs_card.dart';
 import 'package:catalyst_voices/pages/dev_tools/widgets/x_close_button.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_blocs/catalyst_voices_blocs.dart';
